Updated order of actions!
the make_json() method must precede the write_json() method. 
Similarly the make_json() method must follow the put_in_list() method. 

Order of methods called: 
1. if_file_exists 
2. make_json
3. write_json
4. put_in_list
5. make_json
6. write_json
7. load_json
8. print_list 

Now the json text document is updated each time the program is run, and if the file does not yet exist, it is created with the write function after being checked for with the if_file_exists. function.